Add files via upload
This Arduino Sketch aims to provide an interface solution to Continental Corporation UNI NOx Sensor.
The sensor is capable of reading both the O2 [%] and the NOx [ppm].
On cold switch ON, the sensor requires few minutes to warm up. During this time, the serial monitor
shows "Sensor NOT OK ... wait". Once the sensor warms up, the value of O2 [%] and NOx [ppm] settles down
to the actual values and the serial monitor shows "Sensor OK".
Sensor OK => 
- Sensor Supply Status: Supply in Range
- Sensor Temperature Status: At Temperature
- NOx Signal Status: Valid
- O2 Signal Status: Valid
- Heater Mode Status: Auto Mode
- Heater Circuit Status: No Error
- NOx Circuit Status: No Error
- O2 Circuit Status: No Error